+++ src/EtcdTerminal/Configuration/AppSettings.cs
@@ -1,6 +1,6 @@
 namespace EtcdTerminal.Configuration;
 
-public static class EtcdTerminalSettings
+public static class AppSettings
 {
 	public static int PageSize { get; set; } = 30;
 	public static bool TrimInputValues { get; set; } = true;