+++ src/EtcdTerminal/IAppEnvironment.cs
@@ -4,6 +4,5 @@ public interface IAppEnvironment
 {
 	string ConfigDirectoryPath { get; }
 	string ConfigFilePath { get; }
-	string SettingsFilePath { get; }
 	string KeyFilePath { get; }
 }